Add asm files for using a dynamic library
`hello-dl.asm` uses the extern declaration from `msg.asm`.
We can link them statically or create a dynamic library from `msg.asm`
to show the usage.

It cannot be run directly because the `libmsg.so` is not in a default
search path. It needs to be added to the `RPATH` of the
executable that is the same directory as the executable.

`elk` cannot run it either because it does not map the required
dynamic libraries into memory.